+++ flask_pluginkit/__init__.py
@@ -10,13 +10,14 @@
 """
 
 from .pluginkit import PluginManager, push_dcp
-from .utils import Flask, LocalStorage, RedisStorage, MongoStorage, JsonResponse
+from .utils import Flask, LocalStorage, RedisStorage, MongoStorage, \
+    JsonResponse
 from ._installer import PluginInstaller
 from ._web import blueprint
 
 __author__ = "staugur <staugur@saintic.com>"
 
-__version__ = "3.4.0"
+__version__ = "3.4.1"
 
 __all__ = [
     "Flask",